refactor: flatten the walker mod
