--- src/workspace/walker/mod.rs
@@ -1,8 +1,0 @@
-// Copyright 2024 Nelson Dominguez
-// SPDX-License-Identifier: MIT OR Apache-2.0
-
-mod builder;
-mod walk;
-
-pub use builder::*;
-pub use walk::{Walk, WalkResult};
